Change flask/pygeoapi endpoints to Django REST framework.
