--- flask/requirements.txt
@@ -1,8 +1,0 @@
-pandas
-psycopg2
-flask
-pygeoapi
-flask-restful
-datetime
-flask-cors
-openpyxl
